reordering collections in one column functional

+++ server/db/models/Collection.js
@@ -13,7 +13,7 @@ const Collection = db.define("collection", {
     defaultValue: true,
   },
   order: {
-    type: Sequelize.INTEGER,
+    type: Sequelize.REAL,
   },
   category: {
     type: Sequelize.STRING,
